fix feeProtocol

+++ contracts/uniswap_v3/GetUniswapV3PoolDataBatchRequest.sol
@@ -21,7 +21,7 @@ interface IUniswapV3Pool {
             uint16 observationIndex,
             uint16 observationCardinality,
             uint16 observationCardinalityNext,
-            uint8 feeProtocol,
+            uint32 feeProtocol,
             bool unlocked
         );
